Added shop backend

+++ src/main/java/eu/skypotion/shop/buyable/Buyable.java
@@ -2,6 +2,6 @@ package eu.skypotion.shop.buyable;
 
 public interface Buyable {
 
-    long getPrice();
+    double getPrice();
 
 }
